+++ CHANGES.md
@@ -1,7 +1,7 @@
 videojs-record changelog
 ========================
 
-3.9.0 - unreleased
+3.9.0 - 2019/12/29
 ------------------
 
 - New options for image-only mode to change the data type of the